+++ tests/Rust/tests/src/TypeTests.fs
@@ -84,6 +84,23 @@ type TestTypeAttached(a1, a2, a3) =
         with get (i) = arr.[i]
         and set (i) (v) = arr.[i] <- v
 
+type ITestProps =
+    abstract Value1: float with get, set
+    abstract Value: int -> float with get, set
+    abstract Item: int -> float with get, set
+
+type TestProps(arr: float[]) =
+    interface ITestProps with
+        member _.Value1
+            with get () = arr.[1]
+            and set (v) = arr.[1] <- v
+        member _.Value
+            with get (i) = arr.[i]
+            and set (i) (v) = arr.[i] <- v
+        member _.Item
+            with get (i) = arr.[i]
+            and set (i) (v) = arr.[i] <- v
+
 type A  = { thing: int } with
     member x.show() = string x.thing
     static member show (x: A) = "Static: " + (string x.thing)
@@ -657,6 +674,19 @@ let ``Getter and Setter with indexer work`` () =
 [<Fact>]
 let ``Attached Getters Setters and Indexers work`` () =
     let t = TestTypeAttached(1, 2, 3)
+    t.Value1 |> equal 2
+    t.Value1 <- 22
+    t.Value1 |> equal 22
+    t.Value(0) |> equal 1
+    t.Value(0) <- 11
+    t.Value(0) |> equal 11
+    t[2] |> equal 3
+    t[2] <- 33
+    t[2] |> equal 33
+
+[<Fact>]
+let ``Interface Getters Setters and Indexers work`` () =
+    let t = TestProps([| 1; 2; 3 |]) :> ITestProps
     t.Value1 |> equal 2
     t.Value1 <- 22
     t.Value1 |> equal 22