+++ src/Fable.Cli/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+
+* [TS] Fixed interface indexers (#3830) (by @ncave)
+
 ## 4.18.0 - 2024-05-23
 
 ### Fixed